chore(design-tokens): централізувати mobile spacing/theme у пакеті design-tokens
