+++ packages/design-tokens/package.json
@@ -6,13 +6,15 @@
   "description": "Shared Sergeant design tokens & Tailwind preset (web + mobile)",
   "files": [
     "tailwind-preset.js",
-    "tokens.js"
+    "tokens.js",
+    "mobile.ts"
   ],
   "main": "./tokens.js",
   "exports": {
     ".": "./tokens.js",
     "./tokens": "./tokens.js",
-    "./tailwind-preset": "./tailwind-preset.js"
+    "./tailwind-preset": "./tailwind-preset.js",
+    "./mobile": "./mobile.ts"
   },
   "scripts": {
     "typecheck": "echo 'design-tokens has no ts sources'",
